% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apple_csv}
\alias{apple_csv}
\title{Apple Output Template Example
This is an example of an output template that reformats the
statement as a CSV.}
\format{
\subsection{\code{apple_csv}}{

A string that contains Jinja syntax placeholders that match
the \code{apple_template}.
}
}
\usage{
apple_csv
}
\description{
Apple Output Template Example
This is an example of an output template that reformats the
statement as a CSV.
}
\references{
\url{https://arxiv.org/pdf/2407.20007}
}
\keyword{datasets}
