% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apple_statement}
\alias{apple_statement}
\title{Apple Rosetta Statement Example
This is an example of a statement that describes the weight of Apple X.}
\format{
\subsection{\code{apple_statement}}{

A string that specifies an object, a quality, a measurement and a unit.
}
}
\usage{
apple_statement
}
\description{
Apple Rosetta Statement Example
This is an example of a statement that describes the weight of Apple X.
}
\references{
\url{https://arxiv.org/pdf/2407.20007}
}
\keyword{datasets}
