% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apple_template}
\alias{apple_template}
\title{Apple Rosetta Statement Example
This is an example of a template to describe the
weight of a particular apple.}
\format{
\code{apple_template}
A string with slots for an object, a quality, a measurement and a unit.
}
\usage{
apple_template
}
\description{
Apple Rosetta Statement Example
This is an example of a template to describe the
weight of a particular apple.
}
\references{
\url{https://arxiv.org/pdf/2407.20007}
}
\keyword{datasets}
