% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_statements.R
\name{df_to_statements}
\alias{df_to_statements}
\title{Change a data frame into a series of statements based on templates.}
\usage{
df_to_statements(df, templates)
}
\arguments{
\item{df}{A data frame with headers that match the slots in a template. The
dataframe must have one column called "TemplateID" that contains the ID of the
template used to interpret the row.}

\item{templates}{A Rosetta Statement template library created by the
initLibrary function.}
}
\value{
A data frame statement library containing the statements and associated templates.
}
\description{
Change a data frame into a series of statements based on templates.
}
\examples{
templates <- init_library()
templates <- add_template(templates, apple_template)
statements <- df_to_statements(data.frame(object = "Apple X",
                              quality = "weight", value="100.2",
                             unit="grams", TemplateID = "1"),templates)
}
