% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_library.r
\name{init_library}
\alias{init_library}
\title{Load a set of Rosetta templates into a dataframe.}
\usage{
init_library(file = NA)
}
\arguments{
\item{file}{A CSV with the headers 'TemplateID' and 'templateText'.
Headers are case-sensitive and required. Other functions
like \code{df_to_statements()} and \code{add_template()}
will expect a data frame in the
format created by this function.}
}
\value{
A dataframe containing either the loaded templates or the proper
headers for use with other functions.
}
\description{
Load a set of Rosetta templates into a dataframe.
}
\examples{
#Initialize an empty library
templates <- init_library()

#Load a CSV library
apple_templates <- init_library(system.file("extdata", "apple_templates.csv",
 package="rosettaR"))
}
