% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initLibrary.r
\name{initLibrary}
\alias{initLibrary}
\title{Load a set of Rosetta templates into a dataframe.}
\usage{
initLibrary(file)
}
\arguments{
\item{file}{A CSV with the headers 'id', 'templateText' and
'metaTemplateID'. Headers are case-sensitive and required. Other functions
like df_to_statements() and add_template() will expect a data frame in the
format created by this function.}
}
\value{
A dataframe containing either the loaded templates or the proper
headers for use with other functions.
}
\description{
Load a set of Rosetta templates into a dataframe.
}
