% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTemplates.R
\name{loadTemplates}
\alias{loadTemplates}
\title{Load a set of Rosetta templates into a dataframe.}
\usage{
loadTemplates(file)
}
\arguments{
\item{file}{A CSV with the headers 'id', 'templateText' and 'metaTemplateID'. Headers are case-sensitive and required.}
}
\value{
A dataframe containing templates for use with other functions.
}
\description{
Load a set of Rosetta templates into a dataframe.
}
