% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_v2.R
\name{rosetta}
\alias{rosetta}
\title{Combine imperfectly matched datasets}
\usage{
rosetta(d, factor_structure, missing_corr = "normal", id_colnames = NULL)
}
\arguments{
\item{d}{A list of dataframes with imperfectly matched features.}

\item{factor_structure}{A named list. The list names are the factor names.}

\item{missing_corr}{'normal'(default) or 'missing'. 'missing' tries to impute unobserved pairwise correlations}

\item{id_colnames}{optional names  of the columns that uniquely identify each row within a dataset (default is NULL)
Each element is a character vector of feature names for the corresponding factor.}
}
\value{
List of dataframes which contain factor scores.
}
\description{
Forms a complete dataset through latent class concatenation of imperfectly
matched dataset features.
}
\examples{
#----------------------------------------------------------------------------
# Rosetta example
#----------------------------------------------------------------------------

# simulate data
d = sim()

# check feature names
lapply(d$missing, names)

# run rosetta
d_rosetta_missing = rosetta(
  d = d$missing,
  factor_structure = list(
    a = c("a_1", "a_2", "a_3"),
    b = c("b_1", "b_2", "b_3"),
    c = c("c_1", "c_2", "c_3")
  ),
  id_colnames = "ID"
)

d_rosetta_complete = rosetta(
  d = d$complete,
  factor_structure = list(
    a = c("a_1", "a_2", "a_3"),
    b = c("b_1", "b_2", "b_3"),
    c = c("c_1", "c_2", "c_3")
  ),
  id_colnames = "ID"
)

}
