% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_format.R
\name{rosetta_format}
\alias{rosetta_format}
\title{Parse a Rosetta Statement}
\usage{
rosetta_format(s, in_template, out_template = "df")
}
\arguments{
\item{s}{The input statement string.}

\item{in_template}{The Rosetta template string.}

\item{out_template}{The output format ('df' for dataframe, 'rdf' for generic Turtle, or a Jinja string).}
}
\description{
Converts a plain language statement into a structured dataframe using a template.
Supports variables \{\{ var \}\} and optional blocks \link{ ... }.
}
