% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_match.R
\name{rosetta_match}
\alias{rosetta_match}
\title{Match Multiple Statements Against a Template Library}
\usage{
rosetta_match(statements, templates)
}
\arguments{
\item{statements}{A character vector of plain language statements, OR a dataframe containing a column of statements.}

\item{templates}{A dataframe created by \code{init_library()}, containing
'TemplateID' and 'templateText' columns.}
}
\value{
A data.frame in long format with columns: statement_id, statement_text, template_id, variable, value.
}
\description{
Iterates through a vector of statements and attempts to match them against
a library of templates. Returns a long-format dataframe.
}
\examples{
\dontrun{
   # Vector Input
   statements <- c("Kitchener is in Canada", "Paris is in France")

   # Dataframe Input (e.g. from read.csv)
   # df <- read.csv("my_statements.csv")

   templates <- init_library(system.file("extdata/apple_templates.csv", package="rosettaR"))
   results <- rosetta_match(statements, templates)
}
}
