% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_shiny.R
\name{rosetta_shiny}
\alias{rosetta_shiny}
\title{Rosetta Statements Shiny interface}
\usage{
rosetta_shiny()
}
\value{
Nothing.
}
\description{
Rosetta Statements Shiny interface
}
\examples{
\dontrun{
rosetta_shiny()
}
}
