% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettaFormat.R
\name{rosettaFormat}
\alias{rosettaFormat}
\title{Change the format of a statement}
\usage{
rosettaFormat(s, t, f = "df")
}
\arguments{
\item{s}{A statement in the form of a string.}

\item{t}{A Rosetta Template to interpret the string.}

\item{f}{The desired output format. Currently "df" works.}
}
\value{
A variable in the desired output format. Generally a dataframe.
}
\description{
Change the format of a statement
}
\examples{
statement <- "Kitchener is located in Canada"
template <- "{{ city }} is located in {{ country }}"
df <- rosettaFormat(statement, template, "df")
}
