% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_validate.R
\name{rosettaValidate}
\alias{rosettaValidate}
\title{Validate a Rosetta Statement using LinkML.}
\usage{
rosettaValidate(statement, template)
}
\arguments{
\item{statement}{A statement sentence string or data frame of statements.}

\item{template}{A template or data frame of templates.}
}
\value{
True if valid, otherwise a LinkML format report.
}
\description{
Validate a Rosetta Statement using LinkML.
}
