% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rosetta_ui.R
\name{run_rosetta_ui}
\alias{run_rosetta_ui}
\title{Launch the RosettaR UI}
\usage{
run_rosetta_ui()
}
\value{
A character string containing the processed text (if out_template is text),
or a dataframe (if out_template is "df").
}
\description{
Opens a Shiny application to interactively test Rosetta templates and validation.
}
