% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.r
\name{sim}
\alias{sim}
\title{Simulate data for rosetta}
\usage{
sim(loading = matrix(c(0.9, 0.8, 0.7, rep(0, 9), 0.6, 0.7, 0.8, rep(0,
  9), 0.8, 0.9, 0.6), ncol = 3), correlation = matrix(c(1, 0.2, 0.4, 0.2,
  1, 0.3, 0.4, 0.3, 1), ncol = 3), factor_structure = list(a = c(1, 2,
  3), b = c(1, 2, 3), c = c(1, 2, 3)), n_rows = 1000, n_datasets = 3,
  seed = NULL)
}
\arguments{
\item{loading}{The measurement model for x. The '\code{fx}' argument as found in \code{psych::\link[psych]{sim.structure}}.}

\item{correlation}{The structure matrix of the latent variables. The '\code{Phi}' argument as found in \code{psych::\link[psych]{sim.structure}}.}

\item{factor_structure}{A named list. The list names are the factor names.
Each element is a character vector of feature names for the corresponding factor.
Should be ordered corresponding to the rows of the '\code{loading}' argument.}

\item{n_rows}{An integer for the number of rows for each independent dataset.}

\item{n_datasets}{An integer for the number of independent datasets.}

\item{seed}{An integer for the seed.}
}
\value{
Returns a list that contains the complete data and missing data.
}
\description{
This function simulates 'complete' and 'missing' datasets. For each independent
dataset, one variable per domain will be set to missing.
}
\examples{
#----------------------------------------------------------------------------
# Data simulation example
#----------------------------------------------------------------------------
# By default, sim() will simulate 3 variables from 3 different domains.
d_sim <- sim()
str(d_sim)
complete_data <- d_sim$complete
missing_data <- d_sim$missing

}
