% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_template.R
\name{add_template}
\alias{add_template}
\title{Add a template string to the library}
\usage{
add_template(library, template, meta_id = NA)
}
\arguments{
\item{library}{Input must be a library dataframe created by initLibrary()}

\item{template}{This should be one string template.}

\item{meta_id}{It can have a meta-template ID.}
}
\value{
A new library data frame containing the template that was added.
}
\description{
Add a template string to the library
}
\examples{
templates <- init_library()
templates <- add_template(templates, apple_template)
}
