% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta.r
\name{rosetta}
\alias{rosetta}
\title{Combine imperfectly matched datasets}
\usage{
rosetta(d, factor_structure)
}
\arguments{
\item{d}{A list of dataframes with imperfectly matched features.}

\item{factor_structure}{A named list. The list names are the factor names.
Each element is a character vector of feature names for the corresponding factor.}
}
\value{
List of dataframes which contain factor scores.
}
\description{
Forms a complete dataset through latent class concatenation of imperfectly
matched dataset features.
}
\examples{
#----------------------------------------------------------------------------
# Rosetta example
#----------------------------------------------------------------------------
library(rosetta)

# simulate data
d <- sim()

# check feature names
lapply(d$missing, names)

# run rosetta
d_rosetta <- rosetta(
  d = d$missing,
  factor_structure = list(
    a = c("a_1", "a_2", "a_3"),
    b = c("b_1", "b_2", "b_3"),
    c = c("c_1", "c_2", "c_3")
  )
)

}
