% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_format.R
\name{rosetta_format}
\alias{rosetta_format}
\title{Parse a Rosetta Statement}
\usage{
rosetta_format(s, in_template, out_template = "df")
}
\arguments{
\item{s}{The input statement string.}

\item{in_template}{The Rosetta template string.}

\item{out_template}{The output format ('df' for dataframe,
'rdf' for generic Turtle, or a Jinja string).}
}
\value{
Either a data frame or the output template with values filled in from
statement and input template.
}
\description{
Converts a plain language statement into a structured dataframe using
a template.
Supports variables \{\{ var \}\} and optional blocks [ ... ].
}
\examples{
# example code
rosetta_format("Apple X weighs 235 grams", "{{ fruit }} weighs {{ value }} {{ unit}}")

}
