% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta_match.R
\name{rosetta_match}
\alias{rosetta_match}
\title{Match Multiple Statements Against a Template Library}
\usage{
rosetta_match(statements, templates)
}
\arguments{
\item{statements}{A character vector of plain language statements.}

\item{templates}{A dataframe created by \code{init_library()}, containing
'TemplateID' and 'templateText' columns.}
}
\value{
A data.frame in long format with columns: statement_id, statement_text, template_id, variable, value.
}
\description{
Iterates through a vector of statements and attempts to match them against
a library of templates. Returns a long-format dataframe.
}
