% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettaFormat.R
\name{rosettaFormat}
\alias{rosettaFormat}
\title{Change the format of a statement}
\usage{
rosettaFormat(s, in_template, out_template = "df")
}
\arguments{
\item{s}{A statement in the form of a string.}

\item{in_template}{A Rosetta Template to interpret the string.}

\item{out_template}{A Jinja template for the desired output. (Optional; if
not provided, a dataframe is returned)}
}
\value{
A string rendered from the Jinja output template and the Rosetta input template.
}
\description{
Change the format of a statement
}
\examples{
statement <- "Kitchener is located in Canada"
in_template <- "{{ city }} is located in {{ country }}"
out_template <- "CITY,COUNTRY,,{{ city }},{{ country }}"
df <- rosettaFormat(statement, in_template, out_template)
}
